carryon/carryon-mvp-ccl/server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
pydantic>=2.5.0
pydantic-settings>=2.1.0
numpy>=1.24.0
scikit-learn>=1.4.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
python-multipart>=0.0.6 